refactor addMembersDialog and deleteMembersDialog
